requirements.txt
mysql-connector-python
google-generativeai
python-dotenv
openai
